+++ roomai/bridge/__init__.py
@@ -1,11 +1,8 @@
 #!/bin/python
-from roomai.bridge.BridgeUtils   import AllBridgePokerCards
+from roomai.bridge.BridgeUtils   import AllBridgePlayingPokerCards
+from roomai.bridge.BridgeUtils   import AllBridgeBiddingPokerCards
 from roomai.bridge.BridgeUtils   import Direction
-from roomai.bridge.BridgeUtils   import contract_suit_to_rank
-from roomai.bridge.BridgeUtils   import contract_point_to_rank
-from roomai.bridge.BridgeUtils   import suit_str_to_rank
-from roomai.bridge.BridgeUtils   import point_str_to_rank
-from roomai.bridge.BridgeUtils   import BridgePokerCard
+from roomai.bridge.BridgeUtils   import BridgeBiddingPokerCard
 from roomai.bridge.BridgePublicState     import BridgePublicState
 from roomai.bridge.BridgePersonState     import BridgePersonState
 from roomai.bridge.BridgePrivateState    import BridgePrivateState
