+++ roomai/texas/__init__.py
@@ -1,5 +1,7 @@
 #!/bin/python
 from roomai.texas.TexasHoldemUtil   import *
 from roomai.texas.TexasHoldemAction import *
-from roomai.texas.TexasHoldemInfo   import *
+from roomai.texas.TexasHoldemPublicState   import TexasHoldemPublicState
+from roomai.texas.TexasHoldemPersonState   import TexasHoldemPersonState
+from roomai.texas.TexasHoldemPrivateState  import TexasHoldemPrivateState
 from roomai.texas.TexasHoldemEnv    import *
