+++ roomai/texas/play.py
@@ -36,18 +36,18 @@ if __name__ == "__main__":
     for i in range(env.num_normal_players):
         players[i].receive_info(infos[i])
         show_info(infos[i])
-    print "\n"
+    print ("\n")
 
 
     while public_state.is_terminal == False:
         turn = public_state.turn
         action = players[turn].take_action()
-        print "%d player take an action (%s)"%(turn,action.key)
+        print ("%d player take an action (%s)"%(turn,action.key))
         infos, public_state, person_states, private_state = env.forward(action)
         show_public(public_state)
         for i in range(env.num_normal_players):
             players[i].receive_info(infos[i])
             show_info(infos[i])
-        print "\n"
+        print ("\n")
 
     print (public_state.scores)
