+++ tests/testDQN.py
@@ -45,8 +45,8 @@ class DQNTester(unittest.TestCase):
         player = ExamplePlayer()
         dqn   = DqnAlgorithm()
         opponents = [roomai.common.RandomPlayer() for i in range(2)]
-        dqn.train(env=env,players = [player] + opponents + [roomai.common.RandomChancePlayer()], params={})
-        dqn.eval( env=env,players = [player] + opponents + [roomai.common.RandomChancePlayer()], params={})
+        dqn.train(env=env, players = [player] + opponents + [roomai.common.RandomPlayerChance()], params={})
+        dqn.eval(env=env, players = [player] + opponents + [roomai.common.RandomPlayerChance()], params={})
 
     def test_sevenking_dqn(self):
         import logging
@@ -55,8 +55,8 @@ class DQNTester(unittest.TestCase):
         player = SevenKingModel_ThreePlayers()
         algo = DqnAlgorithm()
         opponents = [roomai.common.RandomPlayer() for i in range(2)]
-        algo.train(env=env,players = [player] + opponents + [roomai.common.RandomChancePlayer()], params={"num_normal_players": 3, "num_iters":5})
+        algo.train(env=env, players = [player] + opponents + [roomai.common.RandomPlayerChance()], params={"num_normal_players": 3, "num_iters":5})
         opponents = [roomai.common.RandomPlayer() for i in range(2)]
-        scores = algo.eval(players = [player] + opponents + [roomai.common.RandomChancePlayer()], env=env)
+        scores = algo.eval(players = [player] + opponents + [roomai.common.RandomPlayerChance()], env=env)
         print(scores)
