+++ tests_performance/testBridge.py
@@ -34,7 +34,7 @@ class BridgeTester(unittest.TestCase):
 
     def testAGame(self):
         env = roomai.bridge.BridgeEnv()
-        allcards = list(roomai.bridge.AllBridgePokerCards.values())
+        allcards = list(roomai.bridge.AllBridgePlayingPokerCards.values())
         allcards.sort(key = cmp_to_key(roomai.common.PokerCard.compare))
         infos, public_state, person_states, private_state = env.init({"allcards":allcards, "start_turn":0})
         for i in range(4):
@@ -66,7 +66,7 @@ class BridgeTester(unittest.TestCase):
 
     def testAGame1(self):
         env = roomai.bridge.BridgeEnv()
-        allcards = list(roomai.bridge.AllBridgePokerCards.values())
+        allcards = list(roomai.bridge.AllBridgePlayingPokerCards.values())
         allcards.sort(key = cmp_to_key(roomai.common.PokerCard.compare))
         infos, public_state, person_states, private_state = env.init({"allcards":allcards, "start_turn":0})
         for i in range(4):
@@ -108,7 +108,7 @@ if __name__ == "__main__":
     start = time.time()
     for iter in range(1000):
         env = roomai.bridge.BridgeEnv()
-        allcards = list(roomai.bridge.AllBridgePokerCards.values())
+        allcards = list(roomai.bridge.AllBridgePlayingPokerCards.values())
         allcards.sort(key = cmp_to_key(roomai.common.PokerCard.compare))
         infos, public_state, person_states, private_state = env.init({"allcards":allcards, "start_turn":0})
